+++ src/main/java/com/nhnacademy/illuwa/common/external/user/UserApiClient.java
@@ -20,18 +20,17 @@ public interface UserApiClient {
     @GetMapping(value = "/api/members/addresses")
     List<MemberAddressDto> getAddressByMemberId(@RequestHeader("X-USER-ID") Long memberId);
 
-
-    // 순수 금액 전송 (상품 구매시, 적립금 적용에 사용)
-    @PostMapping(value = "/api/members/points/order/earn")
-    MemberSavePointDto sendTotalPrice(@RequestBody TotalRequest request);
+//    // 순수 금액 전송 (상품 구매시, 포인트 적립에 사용)
+//    @PostMapping(value = "/api/members/points/order/earn")
+//    MemberSavePointDto sendTotalPrice(@RequestBody TotalRequest request);
 
     // 반품 가격
     @PostMapping(value = "/api/members/points/order/return")
     Void sendReturnPrice(@RequestBody TotalRequest totalRequest);
 
-    // 사용된 포인트 전송
+/*    // 사용된 포인트 전송
     @PostMapping(value = "/api/members/points/order/use")
-    MemberUsedPointDto sendUsedPointByMemberId(@RequestBody PointRequest request);
+    MemberUsedPointDto sendUsedPointByMemberId(@RequestBody PointRequest request);*/
 
     // 해당 멤버의 포인트 조회 (RequestHeader 로 memberId 전달)
     @GetMapping(value = "/api/members/points")
