--- src/main/java/com/nhnacademy/illuwa/common/external/user/dto/MemberSavePointDto.java
@@ -1,14 +1,0 @@
-package com.nhnacademy.illuwa.common.external.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class MemberSavePointDto {
-    private BigDecimal totalPrice;
-}
